requirements.txt
numpy
pandas
ipython
pyyaml
pycommand
matplotlib
Sphinx
jinja2
